move test functions around, rename test files with test_
